services/projectr/requirements.txt
Flask==3.0.0
gunicorn==20.1.0
rpy2==3.5.1 # 3.5.2 and up gives errors with rpy2py and py2rpy
pandas==1.4.1
google-cloud-logging